chat-service/app/requirements.txt
fastapi
uvicorn[standard]
redis
numpy
python-dotenv
supabase
aio_pika
pika
requests
google-genai
hnswlib
asyncio